--- app/term/models/__init__.py
@@ -1,8 +1,0 @@
-from .ark import Ark
-from .relationship import Relationship
-from .tag import Tag
-from .term import (Comment, Term, Vote, term_deleted, term_saved, term_tracked,
-                   term_updated)
-from .termset import TermSet
-from .track import Track
-from .association_tables import termset_relationships
