--- app/term/views/__init__.py
@@ -1,5 +1,0 @@
-from app.term.views.api_routes import *
-from app.term.views.tag_views import *
-from app.term.views.relationship_views import *
-from app.term.views.termset_views import *
-from app.term.views.term_views import *
